StudentManagementAttendance/src/test/java/za/ac/cput/StudentManagementAttendance/repository/HODRepositoryImplTest.java
package za.ac.cput.StudentManagementAttendance.repository;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import za.ac.cput.StudentManagementAttendance.domain.HOD;
import za.ac.cput.StudentManagementAttendance.domain.Student;
import za.ac.cput.StudentManagementAttendance.repository.impl.HODRepositoryImpl;
import za.ac.cput.StudentManagementAttendance.repository.impl.StudentRepositoryImpl;
import za.ac.cput.StudentManagementAttendance.repository.repointerfaces.HODRepository;
import za.ac.cput.StudentManagementAttendance.repository.repointerfaces.StudentRepository;

import java.util.Set;

public class HODRepositoryImplTest {

    private HODRepository repository;

    @Before
    public void setUp() throws Exception {
        this.repository = HODRepositoryImpl.getRepository();
    }

    @Test
    public void create() {
        //create a student, add it to the repository
        this.repository.create(null);
        Assert.assertEquals(null, null);
    }

    @Test
    public void read() {
    }

    @Test
    public void update() {
    }

    @Test
    public void delete() {
    }

    @Test
    public void getAll() {
        Set<HOD> students = this.repository.getAll();
        Assert.assertEquals(1, students.size());
    }
}
